# Implementation of function av1_k_means_dim1, av1_k_means_dim2 for AVX2