--- a/Source/Lib/Common/Codec/k_means_template.h
+++ b/Source/Lib/Common/Codec/k_means_template.h
@@ -20,7 +20,7 @@
 #define RENAME_(x, y) AV1_K_MEANS_RENAME(x, y)
 #define RENAME(x) RENAME_(x, AV1_K_MEANS_DIM)
 
-static int RENAME(calc_dist)(const int *p1, const int *p2) {
+static INLINE int RENAME(calc_dist)(const int *p1, const int *p2) {
   int dist = 0;
   for (int i = 0; i < AV1_K_MEANS_DIM; ++i) {
     const int diff = p1[i] - p2[i];
@@ -29,7 +29,7 @@ static int RENAME(calc_dist)(const int *p1, const int *p2) {
   return dist;
 }
 
-void RENAME(av1_calc_indices)(const int *data, const int *centroids,
+INLINE void RENAME(av1_calc_indices)(const int *data, const int *centroids,
                               uint8_t *indices, int n, int k) {
   for (int i = 0; i < n; ++i) {
     int min_dist = RENAME(calc_dist)(data + i * AV1_K_MEANS_DIM, centroids);
@@ -45,7 +45,7 @@ void RENAME(av1_calc_indices)(const int *data, const int *centroids,
   }
 }
 
-static void RENAME(calc_centroids)(const int *data, int *centroids,
+static INLINE void RENAME(calc_centroids)(const int *data, int *centroids,
                                    const uint8_t *indices, int n, int k) {
   int i, j;
   int count[PALETTE_MAX_SIZE] = { 0 };
@@ -76,7 +76,7 @@ static void RENAME(calc_centroids)(const int *data, int *centroids,
   }
 }
 
-static int64_t RENAME(calc_total_dist)(const int *data, const int *centroids,
+static INLINE int64_t RENAME(calc_total_dist)(const int *data, const int *centroids,
                                        const uint8_t *indices, int n, int k) {
   int64_t dist = 0;
   (void)k;