… test/PaletteModeUtilTest.cc