Add AVX2 implementation for kernels av1_k_means_dim1(), av1_k_means_dim2() and similar kernels 
av1_calc_indices_dim1() av1_calc_indices_dim2().

Based on command line below, we have 23% performance improvement.
 ```-enable-altrefs 0 -pred-struct 0 -bit-depth 8 -b out.ivf  -enc-mode 8 -w 1920 -h 1080 -i 1080p.yuv -palette 1 -scm 1 ```