@grill05 2019-12-09
> Based on command line below, we have 23% performance improvement.

@spawlows  Just to clarify, the 23% improvement is in the total run-time of that command, or just in the time spent in the kernel ?