@spawlows 2019-12-10
Performance change for total run-time, on Windows:

Mode | Master FPS | Changes FPS | speed %
-- | -- | -- | --
0 | 0.141 | 0.149 | 5.4
1 | 0.332 | 0.3575 | 7.1
2 | 0.449 | 0.481 | 6.7
3 | 0.985 | 1.121 | 12.1
4 | 2.031 | 2.28 | 10.9
5 | 2.387 | 2.7425 | 13.0
6 | 2.768 | 3.3315 | 16.9
7 | 3.535 | 4.4135 | 19.9
8 | 4.061 | 5.33 | 23.8

This changes work only with params: ```-palette 1 -scm 1```

On Linux we do not see any difference in execution time. From speed test on Windows and Linux kernels work few times faster, so this still reduce CPU consumption. I suppose that on Linux there are difference in threads scheduling and calculation of those kernels are not bottleneck.