@spawlows 2019-12-10
When i run this with one core ```-lp 1``` situation is similar, on Windows we have speed benefits, on linux not. So probably this is some differences between VS and gcc