@spawlows 2019-12-11
Performance change for total run-time, on Windows, and Linux:

Mode | Master FPS | Changes FPS | speed %
-- | -- | -- | --
0 | 0.141 | 0.149 | 5.4
1 | 0.332 | 0.3575 | 7.1
2 | 0.449 | 0.481 | 6.7
3 | 0.985 | 1.121 | 12.1
4 | 2.031 | 2.28 | 10.9
5 | 2.387 | 2.7425 | 13.0
6 | 2.768 | 3.3315 | 16.9
7 | 3.535 | 4.4135 | 19.9
8 | 4.061 | 5.33 | 23.8



After investigation why there is no improvement on linux, I find out that I wrongly executed tests.
For clarification:
"SvtAv1EncApp" use library from build directory even if you copy all the files to another one. So when I build with and without the changes (and copy binaries to another directory), encoder loaded  "libSvtAv1Enc.so.0" from build directory. To fix this issue i would recommend to remove files in Bin/(Release/Debug) before tests. "SvtAv1EncApp" is looking for "libSvtAv1Enc.so.0" in previous build path and then in current path. That might be centos-related issue.